Add tests for functions.hpp, add main.cpp in tests
Continue the unit testing process, and include a main.cpp file in the
test target, which just define the test module and so create the main
function. This was previously done by ApplicationTests.cpp, but this
file was removed because the class Application is not intented to be
unit-tested. And a main file that is not related with any class is more
consistent.