Remove imagesPath and define resourcesPath.
These variables were/are defined in the class Parameters. A path to the
resources directory is more consistent because many classes do use not
only images, but also lang files, config file and so on. And that allows
to move the resources directory, thus allowing the main executable to be
in another place in the file system.