Improve LangManager interface and rename Direction
The public interface of the class LangManager has improved, it now
provides a getter for the default lang (it makes no sense to set the
default lang, because all use of the tr method is done with string
litterals), and some static function for managing available languages.

The elements Top and Bottom in the enum Direction has been renamed to
respectively Up and Down, wich are more consistent.