Use std::string instead of Lang enumeration
This commit changes the public interface of LangManager to use string
for handling lang identifiers, instead of the enumeration Lang.
The main goal of this modification is allowing dynamically added langs,
and make a simpler interface. Because the most of the uses of the
LangManager's setters was was to convert the enum to a string then.
And now a new lang can be added in the resources/lang directory, and it
will immediately available at runtime, without code modifications.