Add a comment explaining an implementation choice
This comment explain why perfect forwarding is not needed in
StateStack::pushState. The trick to maintain a pending list is to store
lambdas functions that remember the arguments to pass to the state
constructor. So perfect forwarding makes no sense, because if just
rvalue-reference are passed to constructors, the underlying object are
dead for a long time when the constructor is effectively called,
resulting in a segfault. So argument are copied by value in the lambda.