Replace enumerated bitfield by use of FlagSet
All uses of the couple unsigned int/enumeration has been replaced to a
strongly-typed flag set based on enum classes rather than weak enum.

This way the code is clearer about what is going on (there is no more
conversions from enum to unsigned) and no enums can be accidentally
mixed. But now a class must be written to listen to fixture destroyings,
because the user data of fixture are a raw pointer to a bit flag (rather
than having the user data used literraly as bit field, as previously),
and raw pointers need to be destroyed. This can be achieved by
inheriting from b2_DestructionListener.

This commit also fix all mistakes that has been made in the
implementation of FlagSet.