Rewrite the interface of FlagSet and Command
The new interface of FlagSet is much more consistent with the intent of
the class, namely provide a *set* of *flags*, based on an enum.
The old one was based on std::bitset, which must know its size at
compile-time, and which allows bit flipping.
There is no portable and easy way to know the size of an enum at
compile-time, and bit flipping makes no sense for almost all flag types.

Because FlagSet is now based on std::set instead of std::bitset,
its constructor, destructor, assignment operator, and so on, are no
longer trivial. This implies that all these methods are correspondly
implicitely deleted for the anonymous union inside the Command struct.
So these methods needed to be explicitely written in Command.
I used this opportunity to encapsulate the attributes, making Command
type-safe and implementation-dependant interfaced.