Optimize the AnimationsManager by using a heap
The old data structure inside the AnimationsManager was a map indexed by
the identifier of the animation. But the most critical method of
AnimationManager, update(), have to get the most important animation,
not an animation by its identifier. So a heap is useful, by giving
always the most important animation at the top of the heap.

In order to keep the heap valid, getters that can possibly invalidate
the heap have now a linear complexity over the number of animation,
rather than logarithmic.
But these operations are called not so often, and update(), which is
called every frame, has now a constant complexity, and a smaller overhead
on data structure operations.

The changes in SpriteSheetAnimation is that the heap structure in
AnimationsManager require the animation to be move-assignable and move-
constructible.