Avoid implicit casts in scripts
The function cast taht was used to implicitely cast values to the right
type in an expression are not longer used. The reason is that these
casts complicated the expression evaluation and are not currently
needed in the scripts. If, in the future, the need of casting variable
arises, an explicit function can be written.