Redirect std::cout and std::cerr to stdout too
Previously the strings that are send to std::cout and std::cerr were
redirected to the in-game console. Now the strings are send to stdout
and stderr too.

The reason is that in case of crash, the error message is impossible to
read because the whole game window is closed.