Flush cout and cerr to stdout and stderr at exit
The std::cout ant std::cerr streams are now redirected to stdout and
stderr at the destruction of the DebugDraw instead of at each frame.

The reason is that there is not simple way to redirect std::cout to both
stdout and the in-game console, so this was made by buffer switching at
each output. This is not very elegant, and when the program crash
without calling DebugDraw::update before exiting (this is usually the
case) nothing were printed on stdout/stderr.

Now the whole content of the console is send to stdout/stderr at the
destruction of the DebugDraw.