Put the default view size as Parameters member
The size of an in-game screen was hardcoded in some places of the code,
this is now fixed.

The information about the size of the level that was stored in
Application and in ScrollingSystem was a sf::IntRect, and now is a
sf::FloatRect. Level coordinates are floating-point, not integer.