Fix the background texture length bug
The texture of the plans had wrong size and were too short, resulting of
the end of the level clipped. This is now fixed, and the various
variables for this computing are a bit more consistent (float and int
have to be mixed, resulting in a mess of static_cast).