Remove Application::initContext
This function was used to set global pointers to useful instances of
objects in the class Context, but a better way to initialize these
pointer has been found, by making the Context class friend of
Application and instanciating a Context object in the Application
constructor.

This allows some others Application members to safely use the Context
pointers in their constructor, instead of creating a method such as
init() that has to be called by Application after the Context pointers
has been set.