Delete init methods that are no longer needed
This requires a bit more work than just moving init method in the
respective constructor, because a constructor that use the gui need that
the gui is fully initialized. the initialization of the gui is achieved
only when a valid window is set to the gui, and this can only be done in
constructors in our case.