Move handleResize in Application as private method
Now the main handling of setting changes is done by the instance of
Application rather than the ParametersState, and the unique case where
the function handleResize is called outside Application is removed.

The setting of the instances such as the sf::Window if the tgui::Gui
should be changed by the object that own them (the instance of
Application), not random objects in the whole code.