Add the explicit specifier where useful
There was many constructor that took one argument but that were not
specifier as *explicit*, allowing unwanted expressions, such as
assignment from the type of the arguement of the constructor.

This commit also add initializing value for uninitialized attribute of
basic type (such as float). This avoids that in some unexpected cases,
unexpected things arise because these attributes has random values.