Reorganize project directories
This commit mainly deletes the bin/ directory, and moves the resources/
directory in the root of the project.

The project is now easy to build with cmake, and cmake can be used in a
developpement workflow. The build consist to make a build/ dir or
something similar, and run `cmake ..`.
But unlike previously, the resources directory will be copied next to
the final executable automatically.