Move the call to b2World::Step in a better place
The call to the Step() method on the main b2World instance has been
moved a few dozen of lines further in the code of PhysicsSystem.cpp

This makes the changes on the various physics entities that are done in
this system effective in the same frame, rather than in the next frame.
The change is not much noticeable, but it is a small improvement.