Handle dependency between entities in save files
The serializing process now serialize the dependency between two entities,
such as item in an inventory, or different parts of a whole entity.

This cannot be done in a clean way without modifying the header of
serializing methods, because the method have to know the name of the
entity currently serialized.