Move json-related functions in a class
This class is name JsonHelper, and is a set of useful functions when
manipulating json values and (de)serializing json to a file.

Previously a part of these functions were located in serialization.hpp,
and the remaining ones in GameState.hpp. Now all is in one file, and is
grouped in one consistent class. This also helps testing, by making
these methods public.