Use typedef rather than inheritance for ContextAccessor
This commit does not compile, it is a first step to a second solution.

The inheritance design lead to ambiguous overload when a base class and
a derived class both inherit from ContextAccessor, although the
inheritance is private. So after a bit of reflexion, a solution with a
typedef appeared.

We don't use containement because the ContextAccessor has no data, this
is just an access to static members of ContextImpl. So we don't need to
have an object of it, a typedef is sufficient, and it doesn't limit the
encapsulation more than the containement.