Separate the tests for getJointDef from other tests
In order to properly test the nullptr return value when the joint type is
not handled, the testJointDefinition was rewritten, and this implies
that if a test fail for a type of joint, all other types won't be
tested. To fix this issue, all joint types have now a dedicated test
case.
And in order to write the context only once (the bodies, the user data
and so on), a specific test suite with a dedicated fixture have be
written.