Use a mock resource rather than sf::Texture for testing ResourceManager
sf::Texture was impossible to test in Travis CI because the X server is
not initialized on their VMs, so another solution was needed.

The mock resource behave much like any resources because it is fully and
uniquely identified by the filename given in the call of loadFromFile,
and is much faster than a true resource to load.