Fix a segfault caused by concurrent accesses to OpenGL
The segfault was due to the fact that the method GameState::initWorld
sometimes threw events to tell to the LoadingState that the sentence to
show has changed. And the LoadingState was directly changing the
corresponding label. Meanwhile, the main thread could also be doing
graphic things. So, it was possible to get a segfault, because two
methods were doing OpenGL things at the same time.

Maybe this is not the exact reason, but this fix seems to be solid
enough to eliminate the bug.