Create a class for doing AABB queries
The process of querying the b2World instance with an AABB (axis-aligned
bounding box) is repeated in multiple places in the code, so a dedicated
class has been written (which also could do other things of this kind,
related to Box2D) in order to write the code only once.

For now it is used only in ArrowPicker.